+++ App.core.trans/App.core.trans/App.core.trans/Models/TransaccionmobileExtend.cs
@@ -12,7 +12,7 @@ namespace App.core.trans.Models
 		public string NombreCliente { get; set; }
 		public DateTime Fecha { get; set; }
 		public decimal Monto { get; set; }
-		public int Longitud { get; set; }
-		public int Latitud { get; set; }
+		public decimal Longitud { get; set; }
+		public decimal Latitud { get; set; }
 	}
 }
